+++ src/models/requests/account_nfts.rs
@@ -41,7 +41,12 @@ impl<'a> Default for AccountNfts<'a> {
 impl<'a> Model for AccountNfts<'a> {}
 
 impl<'a> AccountNfts<'a> {
-    fn new(account: &'a str, id: Option<&'a str>, limit: Option<u32>, marker: Option<u32>) -> Self {
+    pub fn new(
+        account: &'a str,
+        id: Option<&'a str>,
+        limit: Option<u32>,
+        marker: Option<u32>,
+    ) -> Self {
         Self {
             account,
             id,