+++ src/models/requests/random.rs
@@ -31,7 +31,7 @@ impl<'a> Default for Random<'a> {
 impl<'a> Model for Random<'a> {}
 
 impl<'a> Random<'a> {
-    fn new(id: Option<&'a str>) -> Self {
+    pub fn new(id: Option<&'a str>) -> Self {
         Self {
             id,
             command: RequestMethod::Random,