+++ src/models/requests/submit.rs
@@ -60,7 +60,7 @@ impl<'a> Default for Submit<'a> {
 impl<'a> Model for Submit<'a> {}
 
 impl<'a> Submit<'a> {
-    fn new(tx_blob: &'a str, id: Option<&'a str>, fail_hard: Option<bool>) -> Self {
+    pub fn new(tx_blob: &'a str, id: Option<&'a str>, fail_hard: Option<bool>) -> Self {
         Self {
             tx_blob,
             id,